+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to Rainbow-CLI will be documented in this file.
 - #CRRAINB-1518: Display host information when connected
 - #CRRAINB-1533: Add commands for blocking and unblocking a user
 - #CRRAINB-1601: Add command for creating a new application
+- #CRRAINB-1608: Add command for removing an application
 
 ### [1.35.0] - 2018-01-15
 - #RQRAINB-648: Add mass-provisionning commands